New plugin for weaving into the scala compiler (for debugging puposes).
When launching a second instance of Eclipse, you can enable this plugin to
obtain a dump of the stack every time a method belonging to the `scala.tools.nsc.symtab`
package is accessed outside of the `PresentationCompilerThread`.

Be aware, this plugin will have a major impact on the IDE performances, so you should
enable it only during debugging. In particular, if the presentation compiler gets to
an inconsistent state, chances are that symbols are accessed outside of the Presentation
Compiler Thread, and this plugin can make it a lot easier to locate where the offending
code.

Hopefully, this whole plugin will be entirely useless once we will have the Scala Model.